requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
fuzzywuzzy==0.18.0
hashids==1.3.1
matplotlib==3.5.1
mysql-connector-python==8.0.33
numpy==1.21.5
pandas==1.3.5
python-Levenshtein==0.21.1
PyYAML==6.0
Sastrawi==1.0.1
scikit-learn==1.1.2
scikit-lego==0.6.13
scipy==1.7.3
seaborn==0.11.2
sqlparse==0.4.2
urllib3==1.25.11